Move sign_transaction to wallet_state_manager
